backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.30.6
pydantic>=2.10.0
python-dateutil==2.9.0
dateparser==1.2.0
openai==1.40.0
httpx==0.27.2
python-dotenv==1.0.1

# Optional: install when using LLM_PROVIDER=azure
azure-identity==1.17.1

pytest==8.3.2
pytest-asyncio==0.23.8
